Added Currently Reading Component
This commit implements the "Currently Reading" component, which iteratively
displays a list of book in my currently-reading shelf on goodreads.com

The data is gathered via the goodreads API using the axios library, it's then
parsed and put on a 5 second interval. Each entry will be displayed for five
seconds before moving on to the next, cycling back to the beginning once the
last entry is reached.